pyproject.toml
[tool.poetry]
name = "quantdome"
version = "0.1.0"
description = ""
authors = ["painswor28 <painswor@nd.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.1"
pandas = "^1.5"
alpaca-py = "^0.8.2"
pytest = "^7.4.0"
matplotlib = "^3.7.2"
python-decouple = "^3.8"
pyscopg2 = "^66.0.2"
psycopg2 = "^2.9.6"
kafka-python = "^2.0.2"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
